+++ exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/irrationals/log/logConstructors.kt
