--- exact-numbers/src/main/kotlin/xyz/lbres/exactnumbers/irrationals/pi/PiConstructors.kt
@@ -1,5 +1,0 @@
-package xyz.lbres.exactnumbers.irrationals.pi
-
-fun Pi(): Pi = PiImpl()
-// TODO remove this
-fun Pi(isInverted: Boolean): Pi = PiImpl(isInverted)
